w3c-reports/12092014/efl/runned-tests/html/dom/documents/dom-tree-accessors/document.title-02-expected.txt


PASS document.title with head blown away 
FAIL document.title with head blown away 1 assert_equals: expected "" but got "FAIL"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/dom/documents/dom-tree-accessors/document.title-02.xhtml:22:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/dom/documents/dom-tree-accessors/document.title-02.xhtml:16:5)
FAIL Untitled 2 assert_equals: expected "PASS" but got "FAIL"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/dom/documents/dom-tree-accessors/document.title-02.xhtml:28:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/dom/documents/dom-tree-accessors/document.title-02.xhtml:24:5)
FAIL PASS 3 assert_equals: expected "PASS2" but got "FAIL"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/dom/documents/dom-tree-accessors/document.title-02.xhtml:34:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/dom/documents/dom-tree-accessors/document.title-02.xhtml:30:5)
